Fixed issue
I recently opened the project again to try to fix the issue. I changed the name of the apparently missing resource (I think it was a blue circle image) to the name of another resource (a red circle image) and then checked the Form1.cs Design file. As if by magic, the Windows Form returned to working again. I could not view the entire Windows Form because it likes to resize itself randomly and does not like staying in the size I entered in the Properties panel. Oh well. I will get on to replacing the corrupted picture-boxes with the original images they had before then check if everything works as it did before. Phew.